requirements.txt
cookiecutter-data-science
python-dotenv
beautifulsoup4
pySmartDL
psycopg2
pandas
Jinja2

black
flake8
isort
loguru
mkdocs
pip
python-dotenv
tqdm
typer
-e .
